working on adding address from checkout page

+++ database/migrations/2023_12_10_135425_create_addresses_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->index('cityId');
             $table->foreign('cityId')->references('id')->on('cities')->onDelete('cascade')->onUpdate('cascade');
             $table->text('address');
-            $table->integer('zipcode');
+            $table->bigInteger('zipcode');
             $table->integer('number');
             $table->boolean('primary')->default(false);
             $table->timestamps();
